Make the engine build and render as a Windows PE (#57)
* fix(build): the two portability slips a mingw cross build turns up

* feat(core): one ExecutableDirectory, and look there for the engine archives

* build: let a cross build pack its archives with a host-built rxpack

* fix(render): hand the d3d12 swapchain the window's hwnd instead of fishing it out of sdl3.dll

* fix(render): sm 6.6 dxil on windows, and the water pushes through PUSH_CONSTANTS

* fix(render): repack a misaligned d3d12 upload on the gpu when the source will not map

+++ engine/asset/blend_import.cc
@@ -51,10 +51,12 @@ std::string CacheKey(const fs::path &source, const fs::path &script) {
       fs::last_write_time(script, error).time_since_epoch().count();
   if (error)
     return {};
-  const std::string identity = fs::weakly_canonical(source).string() + ":" +
-                               std::to_string(source_size) + ":" +
-                               std::to_string(source_time) + ":" +
-                               std::to_string(script_time);
+  // file_time_type::rep is not a fixed type across standard libraries (libc++
+  // on windows makes it wider than long long), so name the width here.
+  const std::string identity =
+      fs::weakly_canonical(source).string() + ":" + std::to_string(source_size) + ":" +
+      std::to_string(static_cast<long long>(source_time)) + ":" +
+      std::to_string(static_cast<long long>(script_time));
   char key[17];
   std::snprintf(key, sizeof(key), "%016llx",
                 static_cast<unsigned long long>(MakeAssetId(identity).hash));
